refactor(nav): extract browser-router commit pipeline
Move document commit and page fetch into navigation-commit.ts and
navigation-fetch.ts, and return an explicit bodyStrategy from morphHead.
